# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: ai-net Backend API
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: >
    REST & WebSocket API for the ai-net backend — multi-agent task
    orchestration, agent registry, payment reconciliation, background jobs, and
    system health.


    ---


    ## 🔐 Authentication


    ai-net uses header-based cryptographic authentication schemes:


    ### 1. Wallet Authentication (`WalletAuth`)

    - **Header:** `walletpublickey: <Stellar-or-Wallet-Public-Key>`

    - Used for creating tasks, retrieving task state, and managing owned resources.

    - Example: `walletpublickey: GBZXN7PIRZGNMHGA728XZVOG2GUFIDLAZ6AF2I2MD2OCYTAF2K1K4AAA`


    ### 2. Agent Cryptographic Signature Authentication (`AgentSignatureAuth`)

    - **Headers:**
      - `x-signature: <Base64-or-Hex-Ed25519-Signature>`
      - `x-challenge: <Signed-Challenge-String>`
    - Used for agent de-registration and privileged agent operations.

    - The challenge string must match the challenge issued by the server within the replay window.


    ---


    ## ⚡ Rate Limiting


    All public and authenticated endpoints are rate-limited to ensure system stability.

    Rate limit status is communicated via standard HTTP response headers:


    | Header | Description |

    |---|---|

    | `X-RateLimit-Limit` | Maximum requests allowed within the current time window |

...

...


...


    ---


    ## 📡 Live Task Stream (WebSocket)


...


    ```

...

...


...
servers:
...
components:
  securitySchemes:
    WalletAuth:
      type: apiKey
      in: header
      name: walletpublickey
      description: Stellar wallet public key of the task owner / client (e.g., G...)
    AgentSignatureAuth:
      type: apiKey
      in: header
      name: x-signature
      description: Cryptographic Ed25519 signature of the challenge for agent actions
    AgentChallengeAuth:
      type: apiKey
      in: header
      name: x-challenge
      description: Challenge string that was signed by the agent's keypair
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JSON Web Token for authenticated user sessions
...
  schemas:
...
    RegisterAgentRequest:
      type: object
      required:
        - agentId
        - capabilities
        - pricingXLM
        - endpoint
        - stellarPublicKey
      properties:
        agentId:
          type: string
          minLength: 1
        capabilities:
          type: array
          items:
            type: string
            minLength: 1
        pricingXLM:
          type: number
          minimum: 0
        endpoint:
          type: string
        stellarPublicKey:
          type: string
          pattern: ^G[A-Z2-7]{55}$
        bondAmountXLM:
          type: number
          minimum: 0
      additionalProperties: false
...
    Agent:
      type: object
      required:
        - id
        - capabilities
        - pricingXLM
        - endpoint
        - stellarPublicKey
        - reputationScore
        - lastSeenAt
      properties:
        id:
          type: string
          example: agent_crypto_analyst_01
        capabilities:
          type: array
          items:
            type: string
          example:
            - research
            - market_analysis
            - report
        pricingXLM:
          type: number
          example: 0.25
        endpoint:
          type: string
          format: uri
          example: https://agent-crypto.example.com/api
        stellarPublicKey:
          type: string
          example: GABZXN7PIRZGNMHGA728XZVOG2GUFIDLAZ6AF2I2MD2OCYTAF2K1K4XYZ
        reputationScore:
          type: number
          example: 98.5
        lastSeenAt:
          type: string
          format: date-time
          example: 2026-08-25T17:20:00.000Z
    AgentHeartbeatResponse:
      type: object
      required:
        - status
        - lastSeenAt
      properties:
        status:
          type: string
          example: ok
        lastSeenAt:
          type: string
          format: date-time
          example: 2026-08-25T17:30:00.000Z
...
    UnauthorizedError:
      type: object
      required:
        - error
      properties:
        error:
          oneOf:
            - type: string
              example: Missing or invalid walletpublickey authentication header
            - type: object
              properties:
                code:
                  type: string
                  example: UNAUTHORIZED
                message:
                  type: string
                  example: Invalid refresh token
                correlationId:
                  type: string
                  example: 0afbd903-d97c-48bc-bf26-b430fcb863cd
...
    NotFoundError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Task not found
    RateLimitError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Too many requests. Please try again later.
        retryAfter:
          type: integer
          example: 60
...
security:
  - WalletAuth: []
paths:
...
  /api/agents:
    get:
      summary: List registered agents
      operationId: listAgents
      tags:
        - Agents
      security: []
      parameters:
        - in: query
          name: capability
          schema:
            type: string
          description: Filter agents that support this capability
        - in: query
          name: minReputation
          schema:
            type: number
        - in: query
          name: maxPriceXLM
          schema:
            type: number
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/agents/{id}":
    get:
      summary: Get registered agent by ID
      description: Fetches agent profile, capabilities, reputation score, and status
        by unique agentId.
      operationId: getAgent
      tags:
        - Agents
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Unique agent identifier
          example: agent_crypto_analyst_01
      responses:
        "200":
          description: Agent details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
    delete:
      summary: Deregister an agent
      description: >
        Removes an agent from the registry. Requires cryptographic verification
        of an Ed25519 signature generated with the agent's registered Stellar
        secret key.
      tags:
        - Agents
      security:
        - AgentSignatureAuth: []
        - AgentChallengeAuth: []
      operationId: deleteAgent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Unique agent identifier
          example: agent_crypto_analyst_01
        - in: header
          name: x-signature
          required: true
          schema:
            type: string
          description: Base64-encoded Ed25519 signature of the challenge
        - in: header
          name: x-challenge
          required: true
          schema:
            type: string
          description: Plaintext challenge string matching the server challenge
      responses:
        "200":
          description: Agent deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Agent deleted successfully
        "401":
          description: Missing or invalid signature/challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
              example:
                error: Invalid signature
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
...
  /api/agents/register:
    post:
      summary: Register a new specialized agent
      tags:
        - Agents
      security: []
      operationId: registerAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterAgentRequest"
      responses:
        "201":
          description: Agent registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          description: Validation error or Stellar account verification failure
        "429":
          description: Registration rate limit exceeded
  "/api/agents/{id}/heartbeat":
    post:
      summary: Agent heartbeat keep-alive
      description: Updates the agent's lastSeenAt timestamp and keeps its online
        status active.
      tags:
        - Agents
      security: []
      operationId: agentHeartbeat
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: agent_crypto_analyst_01
      responses:
        "200":
          description: Heartbeat recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentHeartbeatResponse"
              example:
                status: ok
                lastSeenAt: 2026-08-25T17:30:00.000Z
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "429":
          description: Heartbeat rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitError"
...